remove dead components
